fix(profile): half a year of experience is half a year, not none
The CV parser already reads a first internship correctly — five months of
work comes out as 0.5 years — and the number was floored to 0 before
anything used it. The floor was deliberate, and it was written to fix a
different bug: 0.5 used to raise on the old int(str(...)) path and become
"unknown", which switches the experience check off entirely. It landed
before the rule that reads a DISTANCE rather than the requirement's own
size, and nobody measured the two together.

They cost exactly one boundary. The door closes at two years of gap, so a
real 0.5 sits one and a half years from an ad asking for two and passes,
while a floored 0 sits a full two and does not. Measured on a real
466-posting archive: 20 offers hidden, sixteen still open, among them a
Junior Data Analyst, an application-consultant role on an ERP suite and a
public-body IT position, all in the user's own city. It hits every new
graduate who uploads a CV.

The fraction now survives into the decision and reads as months where
months are what it is. The CV prompt no longer tells the model to answer 0
under a year, and the chat contract no longer advertises an int.

Two defects found on the way, both in the plan:

An override on these two facts could never be cleared. Emptying the box
sent null, the server read null as "not sent", and the old value stayed —
so a correction typed once outlived every CV uploaded afterwards. Present
-and-null now clears; absent still means untouched, so the chat coach can
keep sending one key at a time. The frontend tells the three cases apart,
because collapsing "unreadable" onto "clear" would let one bad keystroke
wipe a saved fact with a success message on top.

A company bragging about its own age was read as the job's requirement.
The existing guard wants the company introduced BEFORE the number and in
these four the subject comes after it. Widening that search forwards was
measured first and discarded: on 466 postings it fires 29 times and only
twice is it the boast — the rest are real requirements whose sentence
names a company next. Size is the honest test. Every brag says forty; the
largest genuine requirement in the archive is eight.

Verified before/after on the real archive: 20 offers stop being blocked,
zero gain a block, and no flag other than esperienza_richiesta moves.

+++ app/models.py
@@ -159,7 +159,11 @@ class ProfileUpdate(BaseModel):
     # date, a degree it never found — and until now the only way to correct one
     # was to rewrite the whole CV text by hand. They are stored as preferences,
     # not in summary_json, so a correction survives re-uploading the CV.
-    years_experience: int | None = None
+    #: Fractional: a CV with five months of internship is ``0.5``, and an ``int``
+    #: here answered 422 to it — then floored it out of existence on the way in.
+    #: A field PRESENT and null clears the override and hands the fact back to
+    #: the CV; a field left out of the payload leaves the override alone.
+    years_experience: float | None = None
     education_level: str | None = None
     grade: int | None = None
     base_cities: list[str] | None = None
